Should fix when the device is in sleep mode
Hello. 
I have a problem when my laptop is in sleep, after waking up I see an error connecting to the Internet. I found out that when this condition is met,  
my page starts to reload, but since the laptop is in sleep mode, then it does not have access to the Internet, respectively. 

I found a solution to this problem in correcting this part of the code 
When I added setTimeout I stopped getting infinite page reload and further connection error. Now my page reloads when my laptop wakes up.